fix(trusty-agents): satisfy CI's line-cap and test-pointers jobs (not just cargo test/clippy/fmt)
Discovered CI's full job list from .github/workflows/ rather than
assuming cargo test + clippy + fmt was the gate -- it is not. Two jobs
this branch had never been checked against:

1. line-cap (scripts/check_line_cap.sh, 500-SLOC production cap,
   issue #610): the #4115 fix added a `stages`-based health derivation,
   `failed_stages`, and their regression tests directly into
   stores/status.rs, pushing it to 541 SLOC. Split the #[cfg(test)] mod
   out into stores/status_tests.rs (#[path = "status_tests.rs"]),
   mirroring the existing index_feed.rs / index_feed_tests.rs split in
   the same directory. status.rs is back under the cap; the tests
   themselves are unchanged and still pass (52/52 stores:: tests).

2. test-pointers (scripts/check_test_pointers.sh, doc-comment `Test:`
   pointer lint): two PRE-EXISTING dangling citations in
   agent_knowledge.rs, from before this review round --
   `knowledge_route_reports_bound_store_and_tools` and
   `knowledge_route_reports_disabled_endpoint_with_reason` -- neither
   matches the real test names
   (`knowledge_route_reports_bound_store_and_granted_tool`,
   `knowledge_route_reports_disabled_endpoint_never_as_connected`).
   Corrected both citations.

Verified locally: bash scripts/check_line_cap.sh -> "0 violations — OK".
bash scripts/check_test_pointers.sh -> "0 dangling pointers — OK". Also
re-ran bash scripts/check_generation_artifacts.sh and
bash scripts/check_sld.sh (both clean) since this touched tracked .rs
files under their scan. cargo test -p trusty-agents --lib: 2764 passed
(one unrelated pre-existing flake, tools::python_skill's uv/venv race
under concurrent load -- confirmed passing in isolation with
--test-threads=1, untouched by this diff). cargo clippy -p trusty-agents
--lib --tests -- -D warnings: clean. cargo fmt --check: clean.
